script revamp focused on Linux/Mac portability

+++ setup.sh
@@ -34,6 +34,21 @@ function get_script_dir() {
     done
     DIR="$( cd -P "$( dirname "$SOURCE" )" &> /dev/null && pwd )"
     echo "$DIR"
+# Use a combination of readlink and dirname to determine location of script.
+function get_script_dir() {
+    # BASH_SOURCE[0] holds the path to the currently executing script.
+    SOURCE="${BASH_SOURCE[0]}"
+    # Use a loop to handle the case where the script may be a symlink.
+    while [[ -L "$SOURCE" ]]; do 
+        # Use a subshell and `dirname` to resolve the directory containing the symlink.
+        DIR="$( cd -P "$( dirname "$SOURCE" )" &> /dev/null && pwd )"
+        # Resolve symlink using `readlink`.
+        SOURCE="$(readlink "$SOURCE")"
+        # Handle relative symlinks.
+        [[ "$SOURCE" != /* ]] && SOURCE="$DIR/$SOURCE"
+    done
+    DIR="$( cd -P "$( dirname "$SOURCE" )" &> /dev/null && pwd )"
+    echo "$DIR"
 }
 
 # ------ Set Shell Options ------
@@ -75,6 +90,15 @@ fi
 
 # TODO: Add code block here to automatically pull the latest changes of this repository before proceeding.
 
+# ------ Ensure `git` is Installed ------
+
+if ! command -v git &> /dev/null; then 
+    echo "[-] Git is not installed. Installing..."
+    $packageManager install git
+fi
+
+# TODO: Add code block here to automatically pull the latest changes of this repository before proceeding.
+
 # ------ Change Shell in Linux to ZSH ------
 
 if [ "$(check_os)" == "Linux" ]; then
@@ -96,15 +120,57 @@ if [ "$(check_os)" == "Linux" ]; then
     fi
 fi
 
-# ------ Backup Existing Dotfiles ------
+# ------ Backup Shell Dotfiles ------
+
 # TODO: Add functionality to restore original configuration.
 
 backupDir=$(get_script_dir)/.dotfiles.bak
-mkdir backupDir
+if [ ! -d "$backupDir" ]; then
+    mkdir "$backupDir"
+fi
+
+# Backup ~/.bashrc
+if [ -f "$HOME/.bashrc" ]; then
+    cp "$HOME/.bashrc" "$backupDir/.bashrc.bak"
+    echo "[+] Backed up .bashrc"
+else 
+    echo "[-] .bashrc doesn't exist"
+fi
+# Backup ~/.zshrc
+if [ -f "$HOME/.zshrc" ]; then 
+    cp "$HOME/.zshrc" "$backupDir/.zshrc.bak"
+    echo "[+] Backed up .zshrc"
+else 
+    echo "[-] .zshrc doesn't exist"
+fi
+
+# ------ Install Command Line Tools ------
 
 # TODO: Dynamically locate dotfiles to backup.
-cp "$HOME/.zshrc" "$backupDir/.zshrc.bak"
-cp "$HOME/.tmux.conf" "$backupDir/.tmux.conf.bak"
+
+if ! command -v tmux &> /dev/null; then 
+    echo "[-] tmux is not installed. Installing..."
+    $packageManager install tmux
+else 
+    # Backup ~/.tmux.conf
+    if [ -f "$HOME/.tmux.conf" ]; then 
+        cp "$HOME/.tmux.conf" "$backupDir/.tmux.conf.bak"    
+        echo "[+] Backed up .tmux.conf"
+    else 
+        echo "[-] .tmux.conf doesn't exist"
+    fi
+fi 
+
+# TODO: Switch to nvim setup.
+if ! command -v vim &> /dev/null; then
+    echo "[-] vim is not installed. Installing..."
+    $packageManager install vim
+    # TODO: Back up .vimrc.
+fi
+
+
+
+# 
 
 
 # TODO: 
@@ -121,12 +187,20 @@ cp "$HOME/.tmux.conf" "$backupDir/.tmux.conf.bak"
 #     git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 #     # Install zsh-autosuggestions
 #     git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+#     # Install zsh-autosuggestions
+#     git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 
 #     # Install zsh-syntax-highlighting
 #     git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $ZSH_CUSTOM/plugins/zsh-syntax-highlighting
 #     # Install zsh-syntax-highlighting
 #     git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $ZSH_CUSTOM/plugins/zsh-syntax-highlighting
+#     # Install zsh-syntax-highlighting
+#     git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $ZSH_CUSTOM/plugins/zsh-syntax-highlighting
 
+#     # Add plugins to .zshrc
+# else
+#     echo "boo"
+# fi
 #     # Add plugins to .zshrc
 # else
 #     echo "boo"
